### 2025-06-26 #712 @mccroweyclinton-EPA
presubmission request for R package RAQSAPI v2.0.5
Submitting Author Name: Clinton Mccrowey
Submitting Author Github Handle: @mccroweyclinton-EPA
Version submitted: 2.0.5
Repository:  https://github.com/USEPA/RAQSAPI
Submission type: submission

---

Retrieve air monitoring data and associated metadata from the US
  Environmental Protection Agency's Air Quality System service using functions.
  See <https://aqs.epa.gov/aqsweb/documents/data_api.html> for details about
  the US EPA Data Mart API.

RAQSAPI is a R package that was designed specifically as a simple interface to the US Environmental Protection Agency Air Quality System Data Mart API which is a data store of ambient air pollution data collected by EPA, state, local, and tribal air pollution control agencies from over thousands of monitors.  AQS also contains meteorological data, descriptive information about each monitoring station (including its geographic location and its operator), and data quality assurance/quality control information.

Environmental scientists and data scientist

[aqsr](https://github.com/kpkeller/aqsr), RAQSAPI is still under active development and will be updated reflecting future changes to the AQS Data MART API. RAQSAPI is developed by EPA staff. RAQSAPI allow for multi year data pulls. Also RAQSAPI is the only package that I am aware of that is available from CRAN.

YES

RAQSAPI does not have a pkgdown site at this time. There are no HTML man pages hosted online yet.
