### 2025-06-26 #712 @mccroweyclinton-EPA
I need to make a correction, I just searched for my package, "RAQSAPI" in the issues section for this repository looking for this issue and discovered that there is another R package, [epair](https://github.com/ropensci/epair), that appears to work similarly to RAQSAPI, I was totally unaware of this package until now. It looks like epair has already been accepted into ropensci. I am unsure if this code is still being maintain since it has been 3 years since the last commit. I am unsure if I should resend my presubmission request since it seems like there is already another package in ropensci that achieves the same goal as RAQSAPI.
