### 2025-06-30 #712 @mccroweyclinton-EPA
Responding to the issues reported by @ropensci-review-bot

> ## Checks for [RAQSAPI (v2.0.5)](https://github.com/USEPA/RAQSAPI/)
> git hash: [3df1ef76](https://github.com/USEPA/RAQSAPI//tree/3df1ef766c0e9765d9cd0e64add304243b8eff0f)

> * ✖️ does not have a 'contributing' file.

I will consider adding this

> * ✖️ The following function has no documented return value: [aqs_qa_annualpeferomanceeval_by_site]

Is this an error? That function does have a documented (return value)[https://github.com/USEPA/RAQSAPI/blob/3df1ef766c0e9765d9cd0e64add304243b8eff0f/man/aqs_qa_annualperformanceevaltransaction_by_site.Rd].

> * ✖️ These functions do not have examples: [aqs_metadata_service, aqs_services_by_box, aqs_services_by_cbsa, aqs_services_by_county, aqs_services_by_MA, aqs_services_by_pqao, aqs_services_by_site, aqs_services_by_state, aqs_qa_annualpeferomanceeval_by_site].

These functions are intended as internal helper function, not to be called by end users. I have added @internal @roxygen2 tags to these function that I will push as a future commit soon.

> * ✖️ R CMD check found 1 error.

This package is a wrapper around the EPA AQS Datamart API, all functions work without issue, though sometimes a unit test may fail due to connection issues. I am working toward adding mock unit testing to hopefully deal with such issues.
