### 2025-07-03 e30d20a
update the copyright year in LICENSE[,.md] files
