+++ .github/workflows/R-CMD-check.yaml
@@ -2,9 +2,9 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main, httr2, test_github_workflow]
+    branches: [main]
   pull_request:
-    branches: [main, httr2, test_github_workflow]
+    branches: [main]
 
 name: R-CMD-check
 
