### 2025-07-03 0358406
use @internal roxygen2 tags on internal functions, don't include deprecated functions in reference manual
