+++ R/AQSAPI_helperfunctions.R
@@ -24,6 +24,7 @@ server <- "AQSDatamartAPI"
 #' @importFrom magrittr `%<>%`
 #' @importFrom stringr str_detect
 #' @return NULL
+#' @keywords internal
 #' @noRd
 checkaqsparams <- function(...)
 {
@@ -273,6 +274,7 @@ checkaqsparams <- function(...)
 #' @return a string that is properly formatted for use in AQS RESTFUL API
 #'            calls.
 #' @importFrom magrittr `%>%`
+#' @keywords internal
 #' @noRd
 format_variables_for_api <- function(x, separator="&")
 {
@@ -305,6 +307,7 @@ format_variables_for_api <- function(x, separator="&")
 #'                   in the return value.
 #' @return a string that is properly formatted for use in AQS RESTFUL API
 #'            calls.
+#' @keywords internal
 #' @noRd
 format_multiple_params_for_api <- function(x, separator=",")
 {
@@ -337,6 +340,7 @@ format_multiple_params_for_api <- function(x, separator=",")
 #'
 #' @return a string error message that is formatted for httr2 to display
 #'         request errors for end users.
+#' @keywords internal
 #' @noRd
 #'
 #' @example None
@@ -392,7 +396,7 @@ RAQSAPI_error_msg <- function(AQSresponse)
 #' @importFrom dplyr mutate select arrange
 #' @importFrom lubridate ymd_hm
 #' @importFrom glue glue
-#' @importFrom tibble as_tibble
+#' @importFrom tibble tibble
 #' @importFrom rlang caller_call
 #' @importFrom httr2 request req_user_agent req_url_path_append resp_body_json
 #'                   req_perform req_options req_retry req_throttle req_error
@@ -401,6 +405,7 @@ RAQSAPI_error_msg <- function(AQSresponse)
 #'            first item ($Header) is a tibble of header information from the
 #'            AQS API and the second item ($Data) is a tibble of the data
 #'            returned.
+#' @keywords internal
 #' @noRd
 aqs <- function(service, filter = NULL, user = NA,
                     user_key = NA, variables = NULL, AQS_domain = "aqs.epa.gov")
@@ -461,6 +466,7 @@ aqs <- function(service, filter = NULL, user = NA,
 #'         and may not work as expected all the time but overall generally works
 #'         as expected.
 #' @return Boolean
+#' @keywords internal
 #' @noRd
 isValidEmail <- function(email) {
   grepl("\\<[A-Z0-9._%+-]+@[A-Z0-9.-]+\\.[A-Z]{2,}\\>",
@@ -526,6 +532,7 @@ isValidEmail <- function(email) {
 #'            first item ($Header) is a tibble of header information from the
 #'            AQS API and the second item ($Data) is a tibble of the data
 #'            returned.
+#' @keywords internal
 aqs_services_by_site <- function(parameter, bdate, edate,
                                  stateFIPS, countycode, sitenum,
                                  duration = NA_character_, service,
@@ -604,6 +611,7 @@ aqs_services_by_site <- function(parameter, bdate, edate,
 #'            first item ($Header) is a tibble of header information from the
 #'            AQS API and the second item ($Data) is a tibble of the data
 #'            returned.
+#' @keywords internal
 aqs_services_by_county <- function(parameter, bdate, edate,
                                    stateFIPS, countycode, service,
                                    duration = NA_character_,
@@ -677,6 +685,7 @@ aqs_services_by_county <- function(parameter, bdate, edate,
 #'            first item ($Header) is a tibble of header information from the
 #'            AQS API and the second item ($Data) is a tibble of the data
 #'            returned.
+#' @keywords internal
 aqs_services_by_state <- function(parameter, bdate, edate, stateFIPS,
[…]
