+++ R/deprecated.R
@@ -6,8 +6,10 @@
 #'
 #' @family deprecated RAQSAPI functions
 #' @importFrom rlang abort
-#' @export
+#' @noRd
 #' @rdname deprecated
+#' @keywords internal # do not include deprecated function in
+#'                    # package reference manual
 aqs_qa_annualpeferomanceeval_by_site <- function()
 {
   abort(message = "aqs_qa_annualpeferomanceeval_by* functions have been renamed
@@ -25,8 +27,10 @@ aqs_qa_annualpeferomanceeval_by_site <- function()
 #'
 #' @family deprecated RAQSAPI functions
 #' @importFrom rlang abort
-#' @export
+#' @noRd
 #' @rdname deprecated
+#' @keywords internal # do not include deprecated function in
+#'                    # package reference manual
 aqs_qa_annualpeferomanceeval_by_county <- function()
 {
   abort(message = "aqs_qa_annualpeferomanceeval_by* functions have been renamed
@@ -43,8 +47,10 @@ aqs_qa_annualpeferomanceeval_by_county <- function()
 #'
 #' @family deprecated RAQSAPI functions
 #' @importFrom rlang abort
-#' @export
+#' @noRd
 #' @rdname deprecated
+#' @keywords internal # do not include deprecated function in
+#'                    # package reference manual
 aqs_qa_annualpeferomanceeval_by_state <- function()
 {
   abort(message = "aqs_qa_annualpeferomanceeval_by* functions have been renamed
@@ -61,8 +67,10 @@ aqs_qa_annualpeferomanceeval_by_state <- function()
 #'
 #' @family deprecated RAQSAPI functions
 #' @importFrom rlang abort
-#' @export
+#' @noRd
 #' @rdname deprecated
+#' @keywords internal # do not include deprecated function in
+#'                    # package reference manual
 aqs_qa_annualpeferomanceeval_by_pqao <- function()
 {
   abort(message = "aqs_qa_annualpeferomanceeval_by* functions have been renamed
@@ -81,8 +89,10 @@ aqs_qa_annualpeferomanceeval_by_pqao <- function()
 #'
 #' @family deprecated RAQSAPI functions
 #' @importFrom rlang abort
-#' @export
+#' @noRd
 #' @rdname deprecated
+#' @keywords internal # do not include deprecated function in
+#'                    # package reference manual
 aqs_qa_annualpeferomanceeval_by_MA <- function()
 {
    abort(message = "aqs_qa_annualpeferomanceeval_by* functions have been
@@ -99,8 +109,10 @@ aqs_qa_annualpeferomanceeval_by_MA <- function()
 #'
 #' @family deprecated RAQSAPI functions
 #' @importFrom rlang abort
-#' @export
+#' @noRd
 #' @rdname deprecated
+#' @keywords internal # do not include deprecated function in
+#'                    # package reference manual
 aqs_qa_annualpeferomanceevaltransaction_by_site <- function()
 {
   abort(message = "aqs_qa_annualpeferomanceevaltransaction_by* functions have
@@ -117,8 +129,10 @@ aqs_qa_annualpeferomanceevaltransaction_by_site <- function()
 #'
 #' @family deprecated RAQSAPI functions
 #' @importFrom rlang abort
-#' @export
+#' @noRd
 #' @rdname deprecated
+#' @keywords internal # do not include deprecated function in
+#'                    # package reference manual
 aqs_qa_annualpeferomanceevaltransaction_by_county <- function()
 {
   abort(message = "aqs_qa_annualpeferomanceevaltransaction_by* functions have
@@ -135,8 +149,10 @@ aqs_qa_annualpeferomanceevaltransaction_by_county <- function()
 #'
[…]
