### 2025-07-03 62ab465
adjust hexsticker alignment in README
