+++ README.Rmd
@@ -21,7 +21,7 @@ knitr::opts_chunk$set(
 )
 
 ```
-<img src="./vignettes/figures/RAQSAPIhexsticker.png" alt="RAQSAPI hexsticker" width="150" height="150">
+<img src="./vignettes/figures/RAQSAPIhexsticker.png" alt="RAQSAPI hexsticker" align="right" width="150" height="150">
 
 <!-- badges: start -->
   [![Project Status: Active – The project has reached a stable, usable state\
@@ -37,10 +37,10 @@ knitr::opts_chunk$set(
                          https://cran.r-project.org/package=RAQSAPI)
   [![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](
   https://lifecycle.r-lib.org/articles/stages.html)
-  [![license](https://img.shields.io/badge/license-CC0-lightgrey.svg)](
-  https://choosealicense.com/)
-  "[![Last-changedate](https://img.shields.io/badge/last%20change-
-  `r gsub('-', '--', Sys.Date())`-yellowgreen.svg)](/commits/master)"
+  [![license](https://img.shields.io/badge/license-MIT-green)](
+  https://choosealicense.com/licenses/mit/)
+  [![Last-changedate](https://img.shields.io/badge/last%20change-
+  `r gsub('-', '--', Sys.Date())`-yellowgreen.svg)](/commits/master)
   <!-- badges: end -->
 
 ```{r child = "./vignettes/EPA_Disclaimer.Rmd", echo = FALSE, comment = NA}
