### 2025-07-03 73ea8b2
add documentations for the AQS_DATAMARTv2 S3 class
