### 2025-07-03 4505458
prepare RAQSAPI for a version bump, add package:goodpractice back to suggests in DESCRIPTION
