### 2025-07-03 d0396a2
various trivial formating issues, spelling, and grammar improvements, also addresses #27
