+++ vignettes/RAQSAPIvignette.R
@@ -1,5 +1,4 @@
-## ----SETUP, echo = FALSE, message = FALSE, warning = FALSE---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-invisible(library(glue, warn.conflicts = FALSE, quietly = TRUE))
+## ----SETUP, echo = FALSE, message = FALSE, warning = FALSE--------------------
 
 #usethisurl <- "https://usethis.r-lib.org"
 #usethissetup <- "/articles/articles/usethis-setup.html"
@@ -7,7 +6,7 @@ invisible(library(glue, warn.conflicts = FALSE, quietly = TRUE))
 
 #patsetuplink <- glue::glue("{usethisurl}{usethissetup}{usethispatsection}")
 
-## ----RAQSAPIfun_all, echo = FALSE, comment = NA--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+## ----RAQSAPIfun_all, echo = FALSE, comment = NA-------------------------------
 knitr::opts_chunk$set(collapse = TRUE, comment = "#>" )
 invisible(library(magrittr, warn.conflicts = FALSE, quietly = TRUE))
 invisible(library(stringr, warn.conflicts = FALSE, quietly = TRUE))
@@ -25,30 +24,30 @@ invisible(library(glue, warn.conflicts = FALSE, quietly = TRUE))
 # na.omit() %>%
 # as.character()
 RAQSAPI_functions <- c(
-"aqs_annualsummary_by_box",  
-"aqs_annualsummary_by_cbsa",  
-"aqs_annualsummary_by_county",  
+"aqs_annualsummary_by_box",
+"aqs_annualsummary_by_cbsa",
+"aqs_annualsummary_by_county",
 "aqs_annualsummary_by_site",
-"aqs_annualsummary_by_state",  
-"aqs_cbsas",  
-"aqs_classes",  
-"aqs_counties_by_state",  
-"aqs_credentials",  
-"aqs_dailysummary_by_box",  
-"aqs_dailysummary_by_cbsa",  
-"aqs_dailysummary_by_county",  
-"aqs_dailysummary_by_site",  
-"aqs_dailysummary_by_state",  
-"aqs_fields_by_service",  
-"aqs_isavailable",  
-"aqs_knownissues",  
-"aqs_mas",  
-"aqs_monitors_by_box",  
-"aqs_monitors_by_cbsa",  
-"aqs_monitors_by_county",  
-"aqs_monitors_by_site",  
-"aqs_monitors_by_state",  
-"aqs_parameters_by_class",  
+"aqs_annualsummary_by_state",
+"aqs_cbsas",
+"aqs_classes",
+"aqs_counties_by_state",
+"aqs_credentials",
+"aqs_dailysummary_by_box",
+"aqs_dailysummary_by_cbsa",
+"aqs_dailysummary_by_county",
+"aqs_dailysummary_by_site",
+"aqs_dailysummary_by_state",
+"aqs_fields_by_service",
+"aqs_isavailable",
+"aqs_knownissues",
+"aqs_mas",
+"aqs_monitors_by_box",
+"aqs_monitors_by_cbsa",
+"aqs_monitors_by_county",
+"aqs_monitors_by_site",
+"aqs_monitors_by_state",
+"aqs_parameters_by_class",
 "aqs_pqaos",
 "aqs_qa_annualperformanceeeval_by_county",
 "aqs_qa_annualperformanceeval_by_MA",
@@ -60,35 +59,35 @@ RAQSAPI_functions <- c(
[…]
