+++ R/RAQSAPI-package.R
@@ -5,9 +5,9 @@
 #' @title RAQSAPI: A R Interface to The United States Environmental Protection
 #' Agency's Air Quality System Data Mart RESTful API server
 #'
-#' @description RAQSAPI is a package for R that connects the R programming
-#' environment to the United State's Environmental protection agency's
-#' Air Quality System (AQS) Data Mart API for retrieval of air
+#' @description \lifecycle{maturing} RAQSAPI is a package for R that connects
+#' the R programming environment to the United States Environmental protection
+#' agency's Air Quality System (AQS) Data Mart API for retrieval of air
 #' monitoring data.
 #'
 #' There are two things that you must do before using this package.
@@ -15,7 +15,7 @@
 #' 2) Every time this library is reloaded AQS_API_credentials() function
 #'       must be called before continuing.
 #'
-#' please use vignette(RAQSAPI) for more details about this package.
+#' please use \code{vignette(RAQSAPI)} for more details about this package.
 #'
 #' EPA Disclaimer:
 #' This software/application was developed by the U.S. Environmental Protection
@@ -29,6 +29,7 @@
 #' recommendation or favoring by the USEPA. The USEPA seal and logo shall not
 #' be used in any manner to imply endorsement of any commercial product or
 #' activity by the USEPA or the United States Government.
+#'
 #' @docType package
 #' @name RAQSAPI
 #' @keywords internal
