### 2025-07-08 acf086e
Resume using package:goodpractice for package development, update .github/workflows/R-CMD-check.yaml to use updated workflows. add new move .lintr config to project root - modify lintr config, roxygen2: do not include internal functions in family with exported functions, include roxygen2 examples in for internal functions. Fix issues found in aqs_qa_annualperformanceeval_by_county roxygen2 documentation.
