+++ R/deprecated.R
@@ -6,6 +6,13 @@
 #'
 #' @family deprecated RAQSAPI functions
 #' @importFrom rlang abort
+#' @return a tibble or an AQS_Data Mart_APIv2 S3 object of quality assurance
+#'           performance evaluation data. for single monitoring site for the
+#'           sitenum, countycode and stateFIPS requested for the time frame
+#'           between bdate and edate. An AQS_Data_Mart_APIv2 is a 2 item named
+#'           list in which the first item ($Header) is a tibble of header
+#'           information from the AQS API and the second item ($Data) is a
+#'           tibble of the data returned.
 #' @noRd
 #' @rdname deprecated
 #' @keywords internal # do not include deprecated function in
