+++ R/AQSAPI_helperfunctions.R
@@ -481,7 +481,7 @@ isValidEmail <- function(email) {
 #'                 aggregations by site then calls the aqs and returns the
 #'                 result. This helper function is not meant to be called
 #'                 directly from external functions.
-#' @family Aggregate _by_site functions
+#' @family Aggregate _by_site functions AQS_services
 #' @param parameter a character list or a single character string
 #'                    which represents the parameter code of the air
 #'                    pollutant related to the data being requested.
@@ -532,6 +532,29 @@ isValidEmail <- function(email) {
 #'            first item ($Header) is a tibble of header information from the
 #'            AQS API and the second item ($Data) is a tibble of the data
 #'            returned.
+#' @examples # if an user were to call aqs_qa_annualperformanceeval_by_site()
+#'          # for ozone at the Fairhope site in Baldwin County, AL for 2017.
+#'          # [aqs_qa_annualperformanceeval_by_site(parameter = "44201",
+#'          #                                        bdate = as.Date("20170101",
+#'          #                                                format = "%Y%m%d"),
+#'          #                                        edate = as.Date("20171231",
+#'          #                                                 format = "%Y%m%d"),
+#'          #                                        stateFIPS = "01",
+#'          #                                        countycode = "003",
+#'          #                                        sitenum = "0010"
+#'          #                                        )]
+#'          # then aqs_qa_annualperformanceeval_by_site() would call this helper
+#'          # function with the following inputs.
+#'           \dontrun{aqs_services_by_site(parameter = "44201",
+#'                                         bdate=as.Date("20170101",
+#'                                                       format = "%Y%m%d"),
+#'                                         edate=as.Date("20171231",
+#'                                                       format = "%Y%m%d"),
+#'                                         stateFIPS= "01",
+#'                                         countycode = "003",
+#'                                         sitenum = "0010",
+#'                                     service="qaAnnualPerformanceEvaluations")
+#'                   }
 #' @keywords internal
 aqs_services_by_site <- function(parameter, bdate, edate,
                                  stateFIPS, countycode, sitenum,
@@ -611,6 +634,27 @@ aqs_services_by_site <- function(parameter, bdate, edate,
 #'            first item ($Header) is a tibble of header information from the
 #'            AQS API and the second item ($Data) is a tibble of the data
 #'            returned.
+#' @examples # if an user were to call aqs_qa_annualperformanceeval_by_county()
+#'          # for ozone in Baldwin County, AL for 2017.
+#'          # [aqs_qa_annualperformanceeval_by_county(parameter = "44201",
+#'          #                                        bdate = as.Date("20170101",
+#'          #                                                format = "%Y%m%d"),
+#'          #                                        edate = as.Date("20171231",
+#'          #                                                 format = "%Y%m%d"),
+#'          #                                        stateFIPS = "01",
+#'          #                                        countycode = "003"
+#'          #                                        )]
+#'          # then aqs_qa_annualperformanceeval_by_county() would call this
+#'          # helper function with the following inputs.
+#'           \dontrun{aqs_services_by_site(parameter = "44201",
+#'                                         bdate=as.Date("20170101",
+#'                                                       format = "%Y%m%d"),
+#'                                         edate=as.Date("20171231",
+#'                                                       format = "%Y%m%d"),
+#'                                         stateFIPS= "01",
+#'                                         countycode = "003"
+#'                                     service="qaAnnualPerformanceEvaluations")
+#'                   }
 #' @keywords internal
 aqs_services_by_county <- function(parameter, bdate, edate,
[…]
