### 2025-07-09 3eb32bf
remove codecov.yml, fix trival quotation issue in AQSAPI_helperfunctions.R
