### 2025-07-09 d372f60
remove outdated .github/workflows/test-coverage.yaml workflow
