### 2025-07-16 7e41459
use package:lintr to lint code, address code formatting issues found by lintr & fix .lintr config file.
