+++ R/AQSAPI_helperfunctions.R
@@ -37,10 +37,10 @@ checkaqsparams <- function(...)
 
   if ("parameter" %in% names(ellipsis_args))
   {
-    if (nchar(ellipsis_args$parameter) != 5 |
-        !is.character(ellipsis_args$parameter) |
+    if (nchar(ellipsis_args$parameter) != 5 ||
+        !is.character(ellipsis_args$parameter) ||
         !str_detect(ellipsis_args$parameter, "^[:digit:]+$")
-        )
+       )
     {
       error <- TRUE
       errmessage %<>% c("x" =
@@ -51,21 +51,22 @@ checkaqsparams <- function(...)
 
   if ("stateFIPS" %in% names(ellipsis_args))
   {
-    if (nchar(ellipsis_args$stateFIPS) != 2 |
+    if (nchar(ellipsis_args$stateFIPS) != 2 ||
         !is.character(ellipsis_args$stateFIPS)
         )
     {
       error <- TRUE
       errmessage %<>% c("x" =
-      "stateFIPS must be a two digit number (represented as a character string),
-       please pad stateFIPS less than 2 digits with leading zeros"
+          "stateFIPS must be a two digit number (represented as a
+          character string), please pad stateFIPS less than 2 digits with
+          leading zeros"
                         )
     }
   }
 
   if ("countycode" %in% names(ellipsis_args))
   {
-    if (nchar(ellipsis_args$countycode) != 3 |
+    if (nchar(ellipsis_args$countycode) != 3 ||
         !is.character(ellipsis_args$countycode))
     {
       error <- TRUE
@@ -78,7 +79,7 @@ checkaqsparams <- function(...)
 
   if ("sitenum" %in% names(ellipsis_args))
   {
-    if (nchar(ellipsis_args$sitenum) != 4 |
+    if (nchar(ellipsis_args$sitenum) != 4 ||
         !is.character(ellipsis_args$sitenum))
     {
       error <- TRUE
@@ -91,8 +92,8 @@ checkaqsparams <- function(...)
 
   if ("MA_code" %in% names(ellipsis_args))
   {
-    if ((nchar(ellipsis_args$MA_code) != 4 |
-         nchar(ellipsis_args$MA_code) != 3) |
+    if ((nchar(ellipsis_args$MA_code) != 4 ||
+         nchar(ellipsis_args$MA_code) != 3) ||
         !is.character(ellipsis_args$MA_code))
     {
       error <- TRUE
@@ -106,8 +107,8 @@ checkaqsparams <- function(...)
 
   if ("pqao_code" %in% names(ellipsis_args))
   {
-    if ((nchar(ellipsis_args$pqao_code) != 4 |
-         nchar(ellipsis_args$pqao_code) != 3) |
+    if ((nchar(ellipsis_args$pqao_code) != 4 ||
+         nchar(ellipsis_args$pqao_code) != 3) ||
         !is.character(ellipsis_args$pqao_code))
     {
       error <- TRUE
@@ -120,7 +121,7 @@ checkaqsparams <- function(...)
   }
   if ("cbsa_code" %in% names(ellipsis_args))
   {
-    if (nchar(ellipsis_args$cbsa_code) != 5 |
+    if (nchar(ellipsis_args$cbsa_code) != 5 ||
         !is.character(ellipsis_args$cbsa_code))
     {
       error <- TRUE
@@ -132,7 +133,7 @@ checkaqsparams <- function(...)
   }
   if ("POC" %in% names(ellipsis_args))
   {
-    if (nchar(ellipsis_args$POC) != 1 |
+    if (nchar(ellipsis_args$POC) != 1 ||
[…]
