+++ R/bycounty.R
@@ -97,7 +97,7 @@ aqs_monitors_by_county <- function(parameter, bdate, edate, stateFIPS,
 #'                                    bdate = as.Date("20150101",
 #'                                                    format = "%Y%m%d"),
 #'                                    edate=as.Date("20160228",
-#'                                                    format = "%Y%m%d"),
+#'                                                  format = "%Y%m%d"),
 #'                                    stateFIPS = "37",
 #'                                    countycode = "183"
 #'                                    )
@@ -304,7 +304,7 @@ aqs_qa_blanks_by_county <- function(parameter, bdate, edate, stateFIPS,
 #'           #  for Wake County, NC between January and February 2016
 #'  \dontrun{aqs_dailysummary_by_county(parameter = "88101",
 #'                                      bdate = as.Date("20160101",
-#'                                                    format = "%Y%m%d"),
+#'                                                     format = "%Y%m%d"),
 #'                                      edate = as.Date("20170228",
 #'                                                      format = "%Y%m%d"),
 #'                                      stateFIPS = "37",
@@ -571,7 +571,7 @@ aqs_qa_flowrateaudit_by_county <- function(parameter, bdate, edate, stateFIPS,
 #'   \dontrun{aqs_qa_one_point_qc_by_county(parameter= "44201",
 #'                                          bdate = as.Date("20170101",
 #'                                                        format = "%Y%m%d"),
-#'                                          edate = a s.Date("20180131",
+#'                                          edate = as.Date("20180131",
 #'                                                          format = "%Y%m%d"),
 #'                                          stateFIPS = "25",
 #'                                          countycode = "001"
