+++ vignettes/RAQSAPIPurpose.R
@@ -1,5 +1,6 @@
-## ----setup, include = FALSE---------------------------------------------------
-  #`collapse = TRUE,
-  #` comment = "#>"
-  #` )
+## ----setup, include = FALSE--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
 
