+++ R/bycbsa.R
@@ -39,9 +39,9 @@ aqs_monitors_by_cbsa <- function(parameter, bdate, edate, cbsa_code,
 {
   checkaqsparams(parameter, bdate, edate, cbsa_code, cbdate, cedate,
                  return_header)
- # aqs_monitors_by_* functions don't call aqsmultiyearparams() since the
-  #  monitors API call accepts multiple years of data on the server, purrr::pmap
-  #  is used so that the output is consistent with other RAQSAPI functions.
+    #aqs_monitors_by_* functions don't call aqsmultiyearparams() since the
+    #monitors API call accepts multiple years of data on the server, purrr::pmap
+    #is used so that the output is consistent with other RAQSAPI functions.
   params <- tibble(parameter = parameter,
                    bdate = bdate,
                    edate = edate,
