+++ R/zzz.R
@@ -1,14 +1,36 @@
 #nocov start
+#' @section Onload/onunload and on attach functions.
+#' @noRd
+#' @keywords internal # do not include deprecated function in
+#'                    # package reference manual
+
+#' @title .onload
+#' @description setup and configuration of RAQSAPI package when loaded into an
+#' R environment. This function is not to be manually called by the end-user.
+#' @inheritParams base ns-hooks
+#' @return return NULL invisibly
+#' @keywords internal # do not include deprecated function in
+#'                    # package reference manual
+#' @noRd
 .onLoad <- function(libname, pkgname)
 {
   setOldClass("AQS_DATAMART_APIv2")
   utils::globalVariables(names = c("env.RAQSAPI", "AQSObject", "datetime"),
                          package = "RAQSAPI")
-  env.RAQSAPI <- Sys.getenv()
+  env.RAQSAPI <<- Sys.getenv()
   Sys.setenv("_R_CHECK_LENGTH_1_CONDITION_" = "TRUE")
   invisible()
 }
 
+
+#' @title .onUnLoad
+#' @description restore configuration of R environment DURING UNLOADING OF RAQSAPI
+#'           to a state in which is was set before loading RASQSAPI.
+#' @inheritDotParams base ns-hooks
+#' @return return NULL invisibly
+#' @keywords internal # do not include deprecated function in
+#'                    # package reference manual
+#' @noRd
 .onUnLoad <- function(libname, pkgname)
 {
   Sys.setenv(env.RAQSAPI)
@@ -17,9 +39,11 @@
 
 #' @title .onAttach
 #' @description prints out a friendly reminder message to the user when
-#'              the RAQSAPI library is loaded.
+#'              the RAQSAPI library is attached to a R environment.
 #' @inheritParams base .onAttech
 #' @return NULL
+#' @keywords internal # do not include deprecated function in
+#'                    # package reference manual
 #' @noRd
 .onAttach <- function(libname, pkgname)
 {
