+++ vignettes/RAQSAPIPurpose.R
@@ -1,4 +1,4 @@
-## ----setup, include = FALSE--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+## ----setup, include = FALSE------------------------------------------------------------------------------------------------
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
