### 2025-07-18 7a53a97
much needed code cleanup and reformatting with the help of the formatR and lintR packages, line widths change to 125 characters
