+++ R/bycounty.R
@@ -18,36 +18,36 @@
 #'           selected county
 #' @examples # returns an aqs_v2 S3 object containing all SO2 monitors in
 #'           #  Hawaii County, HI that were operating between May 01-02, 2015.
-#'  \dontrun{aqs_monitors_by_county(parameter="42401",
-#'                                  bdate=as.Date("20150501", format="%Y%m%d"),
-#'                                  edate=as.Date("20150502", format="%Y%m%d"),
-#'                                  stateFIPS="15",
-#'                                  countycode="001"
+#'  \dontrun{aqs_monitors_by_county(parameter='42401',
+#'                                  bdate=as.Date('20150501', format='%Y%m%d'),
+#'                                  edate=as.Date('20150502', format='%Y%m%d'),
+#'                                  stateFIPS='15',
+#'                                  countycode='001'
 #'                                  )
 #'          }
 #' @export
-aqs_monitors_by_county <- function(parameter, bdate, edate, stateFIPS,
-                                   countycode, cbdate = NA_Date_,
+aqs_monitors_by_county <- function(parameter, bdate, edate, stateFIPS, countycode, cbdate = NA_Date_,
                                    cedate = NA_Date_, return_header = FALSE)
-{
-  checkaqsparams(parameter, bdate, edate, stateFIPS, countycode,
-                 cbdate, cedate, return_header)
+  {
+  checkaqsparams(parameter, bdate, edate, stateFIPS, countycode, cbdate, cedate, return_header)
 
-  # aqs_monitors_by_* functions don't call aqsmultiyearparams() since the
-  #  monitors API call accepts multiple years of data on the server, purrr::pmap
-  #  is used so that the output is consistent with other RAQSAPI functions.
-  params <- tibble(parameter = parameter,
-                   bdate = bdate,
-                   edate = edate,
-                   stateFIPS = stateFIPS,
-                   countycode = countycode,
-                   service = "monitors",
-                   cbdate = cbdate,
-                   cedate = cedate) %>%
-     dplyr::select_if(function(x) {!all(is.na(x))})
+  # aqs_monitors_by_* functions don't call aqsmultiyearparams() since the monitors API call accepts multiple years of data
+  # on the server, purrr::pmap is used so that the output is consistent with other RAQSAPI functions.
+  params <- tibble(
+    parameter = parameter, bdate = bdate, edate = edate, stateFIPS = stateFIPS,
+    countycode = countycode, service = "monitors", cbdate = cbdate, cedate = cedate
+  ) %>%
+    dplyr::select_if(
+      function(x)
+        {
+        !all(is.na(x))
+      }
+    )
 
   monitors <- purrr::pmap(.l = params, .f = aqs_services_by_county)
-  if (!return_header) monitors %<>% aqs_removeheader
+  if (!return_header)
+    monitors %<>%
+      aqs_removeheader
   return(monitors)
 }
 
@@ -93,39 +93,32 @@ aqs_monitors_by_county <- function(parameter, bdate, edate, stateFIPS,
 #'           returned.
 #' @examples # returns all FRM/FEM PM2.5 data for Wake County, NC between
 #'           #  January 1, 2015 - February 28, 2016
-#'  \dontrun{aqs_sampledata_by_county(parameter = "88101",
-#'                                    bdate = as.Date("20150101",
-#'                                                    format = "%Y%m%d"),
-#'                                    edate=as.Date("20160228",
-#'                                                  format = "%Y%m%d"),
-#'                                    stateFIPS = "37",
-#'                                    countycode = "183"
+#'  \dontrun{aqs_sampledata_by_county(parameter = '88101',
+#'                                    bdate = as.Date('20150101',
+#'                                                    format = '%Y%m%d'),
+#'                                    edate=as.Date('20160228',
[…]
