### 2025-07-22 #712 @mpadge
Thanks @mccroweyclinton-EPA for the progress update. The pkgcheck action is indeed [running as expected](https://github.com/USEPA/RAQSAPI/actions/workflows/pkgcheck.yaml). It is, however, failing for reasons given on the [issue output](https://github.com/USEPA/RAQSAPI/issues/28#issuecomment-3090578932). (And also note that it is set up so to cancel any run as soon as another one starts, hence all the cancelled runs.) Note also that your tests should generally _not_ rely on getting actual results from external APIs (among other reasons, because such tests would likely fail on CRAN), and should instead use some form of mocking, for which see [links in our _Dev Guide_](https://devguide.ropensci.org/pkg_building.html#testing), notably including [our http testing book](https://books.ropensci.org/http-testing/). You should be able to generate mocked results using your API credentials, but then have those results tested in any environments with or without the appropriate credentials.
