### 2025-07-28 3ec3914
minor grammar changes - Thanks Cynthia Stahl, removal of depricated function calls to dplyr::select_if
