+++ vignettes/RAQSAPIFunctions-Long.R
@@ -1,4 +1,4 @@
-## ----RAQSAPIfun_all, echo = FALSE, comment = NA----------------------------------------------------------------------------
+## ----RAQSAPIfun_all, echo = FALSE, comment = NA-------------------------------------------------------------------------------
 knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 invisible(library(magrittr, warn.conflicts = FALSE, quietly = TRUE))
 invisible(library(stringr, warn.conflicts = FALSE, quietly = TRUE))
@@ -106,13 +106,13 @@ RAQSAPI_functions <- c(
 RAQSAPI_functions %>%
   cat(sep = "  \n")
 
-## ----SIGNUPANDCREDENTIALS, echo = FALSE, comment = NA----------------------------------------------------------------------
+## ----SIGNUPANDCREDENTIALS, echo = FALSE, comment = NA-------------------------------------------------------------------------
 signupandcredentials <- paste(".sign_up", ".credentials", sep = "|")
 
 str_subset(string = RAQSAPI_functions, pattern = signupandcredentials) %>%
   cat(sep = "  \n")
 
-## ----METADATAFUNCTIONS, echo = FALSE, comment = NA-------------------------------------------------------------------------
+## ----METADATAFUNCTIONS, echo = FALSE, comment = NA----------------------------------------------------------------------------
 metadatafunctions <- paste(".available",
                            ".fields_by_service",
                            ".knownissues", sep = "|")
@@ -120,7 +120,7 @@ metadatafunctions <- paste(".available",
 str_subset(string = RAQSAPI_functions, pattern = metadatafunctions) %>%
   cat(sep = "  \n")
 
-## ----LISTFUNCTIONS, echo = FALSE, comment = NA-----------------------------------------------------------------------------
+## ----LISTFUNCTIONS, echo = FALSE, comment = NA--------------------------------------------------------------------------------
 listfunctions <- paste(".states",
                        ".counties_by_state",
                        ".sites_by_county",
@@ -134,49 +134,49 @@ listfunctions <- paste(".states",
 str_subset(string = RAQSAPI_functions, pattern = listfunctions) %>%
   cat(sep = "  \n")
 
-## ----_by_Sitefunctions, echo = FALSE, comment = NA-------------------------------------------------------------------------
+## ----_by_Sitefunctions, echo = FALSE, comment = NA----------------------------------------------------------------------------
 by_sitefunctions <- paste("_by_site", sep = '|')
 
 str_subset(string = RAQSAPI_functions, pattern = by_sitefunctions) %>%
   cat(sep = "  \n")
 
-## ----_by_countyfuncions, echo = FALSE, comment = NA------------------------------------------------------------------------
+## ----_by_countyfuncions, echo = FALSE, comment = NA---------------------------------------------------------------------------
 by_countyfunctions <- paste("._by_county", sep = "|")
 
 str_subset(string = RAQSAPI_functions, pattern = by_countyfunctions) %>%
   cat(sep = "  \n")
 
-## ----_by_STATEfunctions, echo = FALSE, comment = NA------------------------------------------------------------------------
+## ----_by_STATEfunctions, echo = FALSE, comment = NA---------------------------------------------------------------------------
 by_STATEfunctions <- paste("._by_state", sep = "|")
 
 str_subset(string = RAQSAPI_functions, pattern = by_STATEfunctions) %>%
   cat(sep = "  \n")
 
-## ----_by_MAfunctions, echo = FALSE, comment = NA---------------------------------------------------------------------------
+## ----_by_MAfunctions, echo = FALSE, comment = NA------------------------------------------------------------------------------
[…]
