### 2025-07-28 932046a
document aqs_sign_up function's return value as recommended by pkgcheck
