### 2025-08-01 38489b2
new .editorconfig, add httptest2 for mocking unit tests, unittest now check for error, PackageCheckArgs: --as-cran in Rproj file, add tidyselect::where to address import issues in multiple functions, update BibTeX file~
