+++ tests/testthat/test-RAQSAPlistfunctions.R
@@ -1,10 +1,10 @@
 #' @importFrom dplyr pull
 #' @importFrom magrittr `%>%`()
 #' @import testthat
+#' @import httptest2
+with_mock_dir("listfunctions",{
 test_that(
   "list functions", {
-    testthat::skip_on_cran()
-    testthat::skip_if_offline()
 
     if (file.exists("local.R"))
       {
@@ -26,32 +26,36 @@ test_that(
     aqs_fields_by_service(service = "list", return_header = TRUE)$Header$status %>%
       expect_match(regexp = "Success")
 
-    aqs_knownissues(return_header = TRUE)$Header$status %>%
-      expect_match(regexp = "Success")
+    aqs_fields_by_service(service = "list", return_header = TRUE) %>%
+      expect_no_failure()
 
-    aqs_revisionhistory(return_header = TRUE)$Header$status %>%
-      expect_match(regexp = "Success")
+    aqs_knownissues(return_header = FALSE) %>%
+      expect_no_failure()
 
-    aqs_states(return_header = TRUE)$Header$status %>%
-      expect_match(regexp = "Success")
+    aqs_revisionhistory(return_header = FALSE) %>%
+      expect_no_failure()
 
-    aqs_counties_by_state(stateFIPS = "40", return_header = TRUE)$Header$status %>%
-      expect_match(regexp = "Success")
+    aqs_states(return_header = FALSE) %>%
+      expect_no_failure()
 
-    aqs_pqaos(return_header = TRUE)$Header$status %>%
-      expect_match(regexp = "Success")
+    aqs_counties_by_state(stateFIPS = "40", return_header = FALSE) %>%
+      expect_no_failure()
 
-    aqs_cbsas(return_header = TRUE)$Header$status %>%
-      expect_match(regexp = "Success")
+    aqs_pqaos(return_header = FALSE) %>%
+      expect_no_failure()
 
-    aqs_classes(return_header = TRUE)$Header$status %>%
-      expect_match(regexp = "Success")
+    aqs_cbsas(return_header = FALSE) %>%
+      expect_no_failure()
 
-    aqs_sites_by_county(stateFIPS = "15", countycode = "001", return_header = TRUE)$Header$status %>%
-      expect_match(regexp = "Success")
+    aqs_classes(return_header = FALSE) %>%
+      expect_no_failure()
 
-    aqs_parameters_by_class(class = "CRITERIA", return_header = TRUE)$Header$status %>%
-      expect_match(regexp = "Success")
+    aqs_sites_by_county(stateFIPS = "15", countycode = "001", return_header = FALSE) %>%
+      expect_no_failure()
+
+    aqs_parameters_by_class(class = "CRITERIA", return_header = FALSE) %>%
+      expect_no_failure()
 
   }
 )
+})
