+++ R/bystate.R
@@ -10,6 +10,7 @@
 #' @inheritParams aqs_services_by_state
 #' @importFrom magrittr `%<>%` `%>%`
 #' @importFrom dplyr select
+#' @importFrom tidyselect where
 #' @param return_header If FALSE (default) only returns data requested.
 #'                        If TRUE returns a AQSAPI_v2 object which is a two item
 #'                        list that contains header information returned from
