### 2025-08-01 e3d51cb
add pkgcheck badge to README, expand checkaqsparams function roxygen2 @return tag
