### 2025-08-05 cfaa3f8
run usethis::use_github_action for ropensci/software-review#712
