### 2025-08-05 fe8633c
fix spelling mistakes, add workflow dispatch: and remove 'master' branch from workflows.
