+++ README.Rmd
@@ -41,7 +41,7 @@ knitr::opts_chunk$set(
   [![Last-changedate](https://img.shields.io/badge/last%20change-
   `r gsub('-', '--', Sys.Date())`-yellowgreen.svg)](/commits/master)
   [![pkgcheck](https://github.com/USEpa/RAQSAPI/workflows/pkgcheck/badge.svg)](https://github.com/USEpa/\
-                RAQSAPI/actions?query=workflow%3Apkgcheck)
+                /RAQSAPI/.github/workflows/pkgcheck.yaml)
   <!-- badges: end -->
 
 ```{r child = "./vignettes/EPA_Disclaimer.Rmd", echo = FALSE, comment = NA}
