### 2025-08-12 #712 @mpadge
Hi @mccroweyclinton-EPA , I'm not sure about that, but suspect that you might see more consistent results by running `testthat::test_local()` or `testthat::test_file()` functions, rather than indirectly via `devtools`. Either way, when mocked tests fail you should see information on the difference between mocked results and those generated by current tests. Failures in different environments generally indicate a need to use the `httptest2` [redacting abilities](https://enpiar.com/httptest2/articles/redacting.html) to redact variable information such as timestamps. Happy to help more if you ping directly from your repo.
