### 2025-08-12 #712 @mccroweyclinton-EPA
@mpage, thank you for the lead. Their does in fact seem to be an issue with how httptest2 is redacting. I will look further.
