+++ vignettes/RAQSAPIPurpose.R
@@ -1,3 +1,6 @@
-## ----setup, include =
-## FALSE------------------------------------------------------------------------------------------------
-knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
+## ----setup, include = FALSE--------------------------------------------------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+
