+++ vignettes/pyaqsapi.R
@@ -1,3 +1,3 @@
-## ----setup,
-## include=FALSE--------------------------------------------------------------------------------------------------
+## ----setup, include=FALSE----------------------------------------------------------------------------------------------------
 knitr::opts_chunk$set(echo = TRUE)
+

[…]
