### 2025-08-22 c9744cb
add test_httptest2 branch to R-CMD-check.yaml GH actions workflow
