+++ .github/workflows/R-CMD-check.yaml
@@ -2,7 +2,9 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main]
+    branches:
+      - main
+      - test_httptest2
   pull_request:
   workflow_dispatch:
 
