+++ .github/workflows/R-CMD-check.yaml
@@ -3,8 +3,7 @@
 on:
   push:
     branches:
-      - main
-      - test_httptest2
+      - '*'
   pull_request:
   workflow_dispatch:
 
