+++ .github/workflows/pkgcheck.yaml
@@ -12,7 +12,7 @@ on:
   push:
     branches:
       - main
-
+      -test_httptest2
 jobs:
   pkgcheck:
     runs-on: ubuntu-latest
