### 2025-08-22 ab16a4f
minor change in HG action workflow pkgcheck.yaml, inst/redact.R now \'replaces\' with redacted instead of \'<redacted>\', use return_header=FALSE in most unit tests
