+++ tests/testthat/test-bystate.R
@@ -26,7 +26,7 @@ test_that(
       bdate = as.Date("20170101", format = "%Y%m%d"),
       edate = as.Date("20171231", format = "%Y%m%d"),
       stateFIPS = "01",
-      return_header = FALSE
+      return_header = TRUE
     ) %>%
       expect_no_error()
 
