+++ inst/httptest2/redact.R
@@ -7,8 +7,8 @@
 # }
 
 
-set_redactor(~ gsub_response(.,getOption(x="aqs_key", default=""), "<redacted>", fixed = TRUE) %>%
-               gsub_response(.,getOption(x="aqs_username", default=""), "<redacted>", fixed = TRUE) %>%
+set_redactor(~ gsub_response(.,getOption(x="aqs_key", default=""), "redacted", fixed = TRUE) %>%
+               gsub_response(.,getOption(x="aqs_username", default=""), "redacted", fixed = TRUE) %>%
                #shorten mock path by removing the base URL
                gsub_response(., "aqs.epa.gov/data/api/", "datamart/", fixed = TRUE) %>%
                #abbreviate long function causes path in mocks

[…]
