+++ .github/workflows/deploy_to_gh-pages-new.yaml
@@ -1,0 +1,36 @@
+name: deploy to github pages
+description: "prepare RAQSAPI vignettes for publishing on github pages"
+
+jobs:
+  build:
+  runs-on: ubuntu-latest
+steps:
+  - uses: actions/checkout@v4
+- name: Set up R
+uses: r-lib/actions/setup-r@v2
+- name: Install dependencies
+run: Rscript -e 'install.packages(c("devtools", "rmarkdown"))'
+- name: Build vignettes
+run: Rscript -e 'devtools::build_vignettes()'
+- name: Copy vignettes to docs
+run: |
+  mkdir -p docs
+cp vignettes/*.html docs/
+  - name: Upload artifact
+uses: actions/upload-pages-artifact@v3
+with:
+  path: docs
+
+deploy:
+  needs: build
+runs-on: ubuntu-latest
+permissions:
+  pages: write
+id-token: write
+environment:
+  name: github-pages
+url: ${{ steps.deployment.outputs.page_url }}
+steps:
+  - name: Deploy to GitHub Pages
+id: deployment
+uses: actions/deploy-pages@v4
