### 2025-09-05 aeb3502
fix formatting in .github/workflows/deploy_to_gh-pages-new.yaml
