### 2025-09-05 474cfe4
modifications to deploy-to_gh-pages-new.yaml to add 'on' and to set permissions
