+++ .github/workflows/deploy_to_gh-pages-new.yaml
@@ -1,6 +1,16 @@
 name: deploy to github pages
 description: "prepare RAQSAPI vignettes for publishing on github pages"
 
+on:
+  push:
+    branches: [main, http2_test]
+  workflow_dispatch:
+
+permissions:
+  contents: read
+  pages: write
+  id-token: write
+
 jobs:
   build:
     runs-on: ubuntu-latest
@@ -24,12 +34,8 @@ jobs:
   deploy:
     needs: build
     runs-on: ubuntu-latest
-    permissions:
-      pages: write
-      id-token: write
     environment:
       name: github-pages
-      url: ${{ steps.deployment.outputs.page_url }}
     steps:
       - name: Deploy to GitHub Pages
         id: deployment
