### 2025-09-16 e7d78dc
change devcontainer to use rocker/geospatial
