### 2025-09-16 37de6ad
removing unneccessary files
