+++ dev/install_RAQSAPI.R
@@ -1,247 +1,0 @@
-# install_RAQSAPI.R This file includes functions to make it easier to deal with building RAQSAPI and is targeted at RAQSAPI
-# development.
-
-library(devtools)
-library(usethis)
-library(dplyr)
-library(testthat)
-library(spelling)
-library(knitr)
-library(magrittr)
-library(glue)
-library(goodpractice)
-library(pkgstats)
-library(pkgcheck)
-library(vroom)
-library(tibble)
-library(desc)
-
-removehttptest2mocks <- function()
-{
-  unlink("./tests/testthat/bybox", recursive = TRUE)
-  unlink("./tests/testthat/bycbsa", recursive = TRUE)
-  unlink("./tests/testthat/byco", recursive = TRUE)
-  unlink("./tests/testthat/byMA", recursive = TRUE)
-  unlink("./tests/testthat/bypqao", recursive = TRUE)
-  unlink("./tests/testthat/bysite", recursive = TRUE)
-  unlink("./tests/testthat/bystate", recursive = TRUE)
-  unlink("./tests/testthat/listfunctions", recursive = TRUE)
-  return(NULL)
-}
-
-#' @name RAQSAPICLEAN
-#' @description removes Package:RASAPI then removes files created during the
-#'   build phase of RAQSAPI including the NAMESPACE file (if NAMESPACE = TRUE
-#'   [default]), compiled vignettes and intermediate files, Rd files and README
-#'   files.
-#' @param NAMESPACE logical, if TRUE (default) deletes the RAQSAPI
-#'   NAMESPACEFILE. Useful if it is expected that building the package will
-#'   alter the NAMESPACE so that the developer can ensure that a new NAMESPACE
-#'   was created.
-#' @note This will remove older versions of RAQSAPI be sure that you intend on
-#'   removing the older version of RAQSAPI before continuing. It has been
-#'   witnessed that sometimes the NAMESPACE is not updated after rebuilding a
-#'   package. By deleting the NAMESPACE file before rebuilding the developer
-#'   can ensure that they are using the most recent version of the NAMESPACE.
-#' @examples
-#'   # To remove RAQSAPI and clean the project directory of files
-#'   #  created during the package build process.
-#'   \dontrun{
-#'            RAQSAPICLEAN()
-#'           }
-#' @internal
-#' @noRd
-RAQSAPICLEAN <- function(NAMESPACE = TRUE)
-  {
-  if ("RAQSAPI" %in% loaded_packages()$package)
-  {
-    unload("RAQSAPI")
-  }
-  if ("RAQSAPI" %in% installed.packages())
-  {
-    remove.packages("RAQSAPI")
-  }
-  if (NAMESPACE == TRUE)
-  {
-    unlink("NAMESPACE")
-  }
-  unlink("README.md")
-  unlink("doc", recursive = TRUE)
-  unlink("man/html", recursive = TRUE)
-  unlink("man", recursive = TRUE)
-  unlink("Meta", recursive = TRUE)
-  unlink("README.html")
-  unlink("README.aux")
-  unlink("README.tex")
-  unlink("README.log")
-  unlink("README.pdf")
-  unlink("index.md")
-  unlink("MD5")
-  unlink("./vignettes/AQSAPI-concordance.tex")
-  unlink("./man/html/", recursive = TRUE)
-  #unlink("./inst/contributorcodeofconduct.md")
-  #unlink(".vignettes/contributing.md")
-  return(NULL)
-}
-
-RAQSAPIinstallMD5sum <- function(path=".")
-{
-  #  tools:::.installMD5sums(pkgDir = ".")  #requires rtools #does not allow excluding files
-
[…]
