+++ dev/contributing.md
@@ -1,37 +1,0 @@
----
-title: "Contributing"
-author:
-- affiliation: |-
-    United States Environmental Protection Agency (US EPA),  
-        Region III,  
-        Air and Radiation Division,  
-        Air Quality and Analysis Branch
-  name: Clinton Mccrowey
-output: github_document
----
-
-# Introduction 
-This document explains how contributions to the RAQSAPI project should be made. 
-
-# License
-All contributions to this repository shall be subject to and adopt the this repository's [License](https://github.com/USEPA/RAQSAPI/blob/main/LICENSE.md). Please make sure that you read and understand the License file before contributing to this project.
-
-# Contributor code of conduct.
-The project's maintainers will enforce this projects's [Contributor Code of Conduct]((https://github.com/USEPA/RAQSAPI/inst/contributorcodeofconduct.md) within the limits that the US
-Environmental Protection Agency policy and federal laws allow.
-
-# all contributions should be accompanied by an issue number
-Be sure to search for an [exiting issue](https://github.com/USEPA/RAQSAPI/issues) that addresses what your contribution is attempting to change. If an existing issue has not already been made create a new issue that briefly explains what changes
-your contribution is attempting to make. Be sure that the issue number is associated with the pull request (PR).
-
-# Ways to contribute
-The RAQSAPI maintainers are glad that you are interested in contributing to this project. You do not have to be programmer to contribute. We value your input on improving this project. If you believe that there is a way to improve RAQSAPI then we want to hear from you. In addition to submitting a pull request for code changes, here are a few other ways in which you can contribute.
-* submit an issue on [RAQSAPI's issue tracker](https://github.com/USEPA/RAQSAPI/issues).
-* contribute new documentation or improve the project's documentation.
-  * fix spelling/grammar errors.
-  * address issues in Roxygen2 tags
-* submit feature requests via the issue tracker.
-* improve unit testing
-
-# before sumbitting
-Before submittin gyour code changes, be sure ot run lintr and 
