+++ dev/contributorcodeofconduct.Rmd
@@ -1,16 +1,0 @@
----
-title: "Contributor Code Of Conduct"
-author:
-- affiliation: |-
-    United States Environmental Protection Agency (US EPA),  
-        Region III,  
-        Air and Radiation Division,  
-        Air Quality and Analysis Branch
-  name: Clinton Mccrowey
-output: github_document
----
-
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = TRUE)
-```
-Please refer to the [GSA TTS handbook](https://handbook.tts.gsa.gov/about-us/code-of-conduct/) for the Code of Conduct
