### 2025-10-07 62a604e
new codeinvenstory.json file, rebuild codemeta.json, remove erronius @section roxygen2 tags, remove inst/schemaorg.json, and many improvements to the projects overall documentation thanks Maria Morresi (ASRC Federal)
