+++ R/setupfunctions.R
@@ -1,6 +1,3 @@
-#' @section  RAQSAPI setup functions
-
-
 #' @title aqs_credentials
 #' @description \lifecycle{stable}
 #'                 Sets the user credentials for the AQS API. This function
@@ -41,13 +38,13 @@ aqs_credentials <- function(username = NA_character_, key = NA_character_)
   {
     warning("Please enter a valid username and key  \n")
   }
-  return(NULL)
+  invisible()
 }  #no cov end
 
 
 #' @title aqs_sign_up
 #' @description \lifecycle{stable}
-#'              Use this service to register as a new user or to reset an
+#'              Use this function to register as a new user or to reset an
 #'              existing user's key. A verification email will be sent to the
 #'              email account specified. To reset a password: If the request is
 #'              made with an email that is already registered, a new key will
@@ -60,7 +57,7 @@ aqs_credentials <- function(username = NA_character_, key = NA_character_)
 #'                 the account specified. Follow the instructions
 #'                 in the verification e-mail before proceeding to use any other
 #'                 functionality of the AQS API. Register your credential
-#'                 with the @3 [aqs_credentials()] before using the
+#'                 with [aqs_credentials()] before using the
 #'                 other functions in this library.
 #' @note The '@' character needs to be escaped with the '/' character.
 #' @importFrom glue glue
@@ -87,5 +84,5 @@ aqs_sign_up <- function(email)
   # user agent string (currently not implemented) on the API %>% req_user_agent(string = user_agent)
   glue("A verification email will be sent to {email}  \n") %>%
     message()
-  return(NULL)
+  invisible()
 }  #nocov end

[…]
