### 2025-10-07 228b3ec
new devcontainer.json and Dockerfile under .decontainer/
