### 2025-12-09 21ba1b5
fixed many lints suggested by R package:lintr, add DOI info to multiple package metadata files, fixed issues with mock files
