+++ vignettes/RAQSAPIvignette.R
@@ -1,121 +1,111 @@
-## ----SETUP, echo = FALSE, message = FALSE, warning = FALSE-------------------------------------------------------------------
+## ----SETUP, echo = FALSE, message = FALSE, warning = FALSE-----------------------------------------------------------------
 invisible(library(glue, warn.conflicts = FALSE, quietly = TRUE))
 
-## ----RAQSAPIfun_all, echo = FALSE, comment = NA------------------------------------------------------------------------------
+## ----RAQSAPIfun_all, echo = FALSE, comment = NA----------------------------------------------------------------------------
 knitr::opts_chunk$set(collapse = TRUE, comment = "#>")
 invisible(library(magrittr, warn.conflicts = FALSE, quietly = TRUE))
 invisible(library(stringr, warn.conflicts = FALSE, quietly = TRUE))
 invisible(library(tibble, warn.conflicts = FALSE, quietly = TRUE))
 invisible(library(glue, warn.conflicts = FALSE, quietly = TRUE))
 
-#AQAD-33, using list.files works when knitting the vignette but does
-#not seem to work when the vignette is knitted as part of a package.
-#Instead function names need to be hard coded.
-# RAQSAPI_functions <- list.files(path = "./man/", pattern = ".Rd$") %>%
-# stringr::str_remove_all(pattern = ".Rd") %>%
-# str_replace("ni_infix_operator", replacement = NA_character_) %>%
-# str_replace("RAQSAPI", replacement = NA_character_) %>%
-# str_replace_all("_services_", replacement = NA_character_) %>%
-# na.omit() %>%
-# as.character()
 RAQSAPI_functions <- c(
-"aqs_annualsummary_by_box",
-"aqs_annualsummary_by_cbsa",
-"aqs_annualsummary_by_county",
-"aqs_annualsummary_by_site",
-"aqs_annualsummary_by_state",
-"aqs_cbsas",
-"aqs_classes",
-"aqs_counties_by_state",
-"aqs_credentials",
-"aqs_dailysummary_by_box",
-"aqs_dailysummary_by_cbsa",
-"aqs_dailysummary_by_county",
-"aqs_dailysummary_by_site",
-"aqs_dailysummary_by_state",
-"aqs_fields_by_service",
-"aqs_isavailable",
-"aqs_knownissues",
-"aqs_mas",
-"aqs_monitors_by_box",
-"aqs_monitors_by_cbsa",
-"aqs_monitors_by_county",
-"aqs_monitors_by_site",
-"aqs_monitors_by_state",
-"aqs_parameters_by_class",
-"aqs_pqaos",
-"aqs_qa_annualperformanceeval_by_county",
-"aqs_qa_annualperformanceeval_by_MA",
-"aqs_qa_annualperformanceeval_by_pqao",
-"aqs_qa_annualperformanceeval_by_site",
-"aqs_qa_annualperformanceeval_by_state",
-"aqs_qa_annualperformanceevaltransaction_by_county",
-"aqs_qa_annualperformanceevaltransaction_by_MA",
-"aqs_qa_annualperformanceevaltransaction_by_pqao",
-"aqs_qa_annualperformanceevaltransaction_by_site",
-"aqs_qa_annualperformanceevaltransaction_by_state",
-"aqs_qa_blanks_by_county",
-"aqs_qa_blanks_by_MA",
-"aqs_qa_blanks_by_pqao",
-"aqs_qa_blanks_by_site",
-"aqs_qa_blanks_by_state",
-"aqs_qa_collocated_assessments_by_county",
-"aqs_qa_collocated_assessments_by_MA",
-"aqs_qa_collocated_assessments_by_pqao",
[…]
