+++ tests/testthat/test-AQS_DATAMART_API-S3-class.R
@@ -27,7 +27,7 @@ test_that(
 #' @importFrom magrittr `%>%`()
 #' @importFrom tibble tibble tribble
 #' @importFrom lubridate mdy_hms now year
-#' @import from glue gluedata:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAABIAAAASCAYAAABWzo5XAAAAbElEQVR4Xs2RQQrAMAgEfZgf7W9LAguybljJpR3wEse5JOL3ZObDb4x1loDhHbBOFU6i2Ddnw2KNiXcdAXygJlwE8OFVBHDgKrLgSInN4WMe9iXiqIVsTMjH7z/GhNTEibOxQswcYIWYOR/zAjBJfiXh3jZ6AAAAAElFTkSuQmCC
+#' @import from glue glue
 #' @import testthat
 test_that(
   "test AQS_DATAMART_APIv2_validator", {
@@ -56,6 +56,6 @@ test_that(
     )
     list(Header = fakeheader, Data = fakeData) %>%
       RAQSAPI:::AQS_DATAMART_APIv2_validator() %>%
-      expect_no_failure()
+      expect_no_error()
   }
 )
