+++ tests/testthat/test-bystate.R
@@ -10,10 +10,10 @@ if (file.exists("local.R"))
   datamartAPI_user <- AQScredentials$datamartAPI_user
   datamartAPI_key <- AQScredentials$datamartAPI_key
 } else
-{
-  datamartAPI_user <- Sys.getenv("RAQSAPIUSERNAME", names = TRUE)
-  datamartAPI_key <- Sys.getenv("RAQSAPIKEY", names = TRUE)
-}
+  {
+    datamartAPI_user <- Sys.getenv("RAQSAPIUSERNAME", names = TRUE)
+    datamartAPI_key <- Sys.getenv("RAQSAPIKEY", names = TRUE)
+  }
 RAQSAPI::aqs_credentials(username = datamartAPI_user, key = datamartAPI_key)
 
 
