+++ vignettes/pyaqsapi.R
@@ -1,3 +1,1 @@
-## ----setup, include=FALSE---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-knitr::opts_chunk$set(echo = TRUE)
-
+## ----setup, include=FALSE-----------------------------------------------------

[…]
