### 2025-12-09 c2f3609
move inst/CODE_OF_CONDUCT.md to root, add contributing.md, rebuild README.md
