+++ tests/testthat/test-bybox.R
@@ -14,70 +14,70 @@ if (file.exists("local.R"))
   }
 RAQSAPI::aqs_credentials(username = datamartAPI_user, key = datamartAPI_key)
 
-with_mock_dir("bybox",{
-test_that(
-  "bybox functions", {
+with_mock_dir("bybox", {
+  test_that(
+    "bybox functions", {
 
-    aqs_sampledata_by_box(
-      parameter = "44201",
-      bdate = as.Date("20150501", format = "%Y%m%d"),
-      edate = as.Date("20150502", format = "%Y%m%d"),
-      minlat = "33.3",
-      maxlat = "33.6",
-      minlon = "-87.0",
-      maxlon = "-86.7",
-      return_header = TRUE
-    ) %>%
-      expect_no_error()
+      aqs_sampledata_by_box(
+        parameter = "44201",
+        bdate = as.Date("20150501", format = "%Y%m%d"),
+        edate = as.Date("20150502", format = "%Y%m%d"),
+        minlat = "33.3",
+        maxlat = "33.6",
+        minlon = "-87.0",
+        maxlon = "-86.7",
+        return_header = TRUE
+      ) %>%
+        expect_no_error()
 
-    aqs_monitors_by_box(
-      parameter = "44201",
-      bdate = as.Date("19950101", format = "%Y%m%d"),
-      edate = as.Date("19951231", format = "%Y%m%d"),
-      minlat = "33.3",
-      maxlat = "33.6",
-      minlon = "-87.0",
-      maxlon = "-86.7",
-      return_header = FALSE
-    ) %>%
-      expect_no_error()
+      aqs_monitors_by_box(
+        parameter = "44201",
+        bdate = as.Date("19950101", format = "%Y%m%d"),
+        edate = as.Date("19951231", format = "%Y%m%d"),
+        minlat = "33.3",
+        maxlat = "33.6",
+        minlon = "-87.0",
+        maxlon = "-86.7",
+        return_header = FALSE
+      ) %>%
+        expect_no_error()
 
-    aqs_annualsummary_by_box(
-      parameter = "44201",
-      bdate = as.Date("20150501", format = "%Y%m%d"),
-      edate = as.Date("20150502", format = "%Y%m%d"),
-      minlat = "33.3",
-      maxlat = "33.6",
-      minlon = "-87.0",
-      maxlon = "-86.7",
-      return_header = FALSE
-    ) %>%
-      expect_no_error()
+      aqs_annualsummary_by_box(
+        parameter = "44201",
+        bdate = as.Date("20150501", format = "%Y%m%d"),
+        edate = as.Date("20150502", format = "%Y%m%d"),
+        minlat = "33.3",
+        maxlat = "33.6",
+        minlon = "-87.0",
+        maxlon = "-86.7",
+        return_header = FALSE
+      ) %>%
+        expect_no_error()
 
-    aqs_dailysummary_by_box(
-      parameter = "44201",
-      bdate = as.Date("20150501", format = "%Y%m%d"),
-      edate = as.Date("20150502", format = "%Y%m%d"),
-      minlat = "33.3",
-      maxlat = "33.6",
-      minlon = "-87.0",
-      maxlon = "-86.7",
-      return_header = FALSE
-    ) %>%
-      expect_no_error()
+      aqs_dailysummary_by_box(
+        parameter = "44201",
+        bdate = as.Date("20150501", format = "%Y%m%d"),
+        edate = as.Date("20150502", format = "%Y%m%d"),
+        minlat = "33.3",
+        maxlat = "33.6",
+        minlon = "-87.0",
+        maxlon = "-86.7",
+        return_header = FALSE
+      ) %>%
[…]
