### 2025-12-12 566ea2c
fixed a lot of lint found by lintr, partially addresses #29
