+++ inst/httptest2/redact.R
@@ -1,3 +1,9 @@
+if (getOption("raqsapi.redactor_installed", FALSE)) {
+  invisible(NULL)
+} else {
+  options(raqsapi.redactor_installed = TRUE)
+}
+
 httptest2::set_redactor(~{
   resp <- .
   key <- getOption("aqs_key")
@@ -10,13 +16,14 @@ httptest2::set_redactor(~{
     resp <- httptest2::gsub_response(resp, as.character(user), "<redacted>", fixed = TRUE)
   }
 
-  #shorten mock path by removing the base URL
+  # shorten mock path by removing the base URL
   resp <- httptest2::gsub_response(resp, "aqs.epa.gov/data/api/", "datamart/", fixed = TRUE)
-  #abbreviate long function causes path in mocks
+  # abbreviate long function causes path in mocks
   resp <- httptest2::gsub_response(resp, "qaAnnualPerformanceEvaluations", "QAape", fixed = TRUE)
-  #abbreviate long function causes path in mocks
+  # abbreviate long function causes path in mocks
   resp <- httptest2::gsub_response(resp, "transactionsQaAnnualPerformanceEvaluations", "tQAape", fixed = TRUE)
   resp <- httptest2::gsub_response(resp, "quarterlyData", "QD", fixed = TRUE)
 
   resp
 })
+
