+++ tests/testthat/test-bysite.R
@@ -16,6 +16,7 @@ if (file.exists("local.R"))
 RAQSAPI::aqs_credentials(username = datamartAPI_user, key = datamartAPI_key)
 
 
+#with_mock_api({
 with_mock_dir("bysite", {
   test_that(
     "bysite functions", {
