+++ tests/testthat/test-bystate.R
@@ -16,7 +16,7 @@ if (file.exists("local.R"))
   }
 RAQSAPI::aqs_credentials(username = datamartAPI_user, key = datamartAPI_key)
 
-
+#with_mock_api({
 with_mock_dir("bystate", {
   test_that(
     "bystate functions", {
