### 2025-12-12 #712 @mccroweyclinton-EPA
@mpadge, @jhollist  and @ldecicco-USGS, I just pushed a commit so that you can see the code that I am working with, the GH action is getting the same error that I am getting locally. I fixed the long path name issue a few months back. This is what I have been spending the past few weeks trying to fix. It works fine it I don't use the mocks provided by httptest2.
